requirements.txt
beautifulsoup4
Mastodon.py
feedparser
